requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
mtcnn==0.1.1
tkinter  # Usually comes with Python